requirements.txt
# Core Big Data Dependencies
pyspark==3.5.1
delta-spark==3.1.0

# FastAPI backend
fastapi==0.111.0
uvicorn[standard]==0.29.0

# YAML Configuration Loader
pyyaml==6.0.1

# For optional data exploration
pandas==2.2.2

# Logging and Utility
loguru==0.7.2

# For testing 
pytest==8.2.2
httpx==0.27.0
